Fixes for recursive_subprograms
Set_Units: New subprogram, needed to set the units later for a
given context.
Use Set_Units, for the lalcheck driver we were ending up with an empty
set of units because of order of initialization and as a result, we
could not get dispatching calls working on recursive_subprograms.

Simplify removal of duplicate handling between CallExpr and BaseId.
Inline check_body function, and ensure the binding of variables works
properly. For some reason (LKQL bug?), the binding of the variable 'b'
was not working properly, leading to missed detection on e.g. R525-005.

Change-Id: Idc07d6bea8abbe1ed266d4a1209c971741deebe9