Factorize matching logic in 'identifier_prefixes' to avoid duplication
Before this commit, we matched on the rule's ode twice: once to know if
the node violates the rule, and the other time to get the diagnostic
message related to the node's kind.
Now we use a function to get from a node, either a diagnostic message
when the node violates the rule, or unit if the node is "valid".