Do not rely on Full_Name when processing rule files
Full_Name is not a canonicalized name of a file, so relying on it to
avoid adding the same rule file several time is not working when the
same rule directory is specified more than once with different path
names (e.g.: ./dir and ./../parent_dir/dir).

Instead, rely on the Virtual_File comparison function which works in
all cases.